+++ incuna_auth/forms.py
@@ -1,8 +1,6 @@
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Div, Field, Layout, Submit
-from django import forms
-from django.contrib.auth.forms import AuthenticationForm, PasswordResetForm
-from django.utils.translation import ugettext_lazy as _
+from django.contrib.auth.forms import PasswordResetForm
 
 
 class CrispyPasswordResetForm(PasswordResetForm):
@@ -13,8 +11,3 @@ class CrispyPasswordResetForm(PasswordResetForm):
             Submit('submit', 'Reset'),
         )
     )
-
-
-class IncunaAuthenticationForm(AuthenticationForm):
-    username = forms.CharField(label=_('Email'), max_length=320, widget=forms.TextInput(attrs={'type': 'email'}))
-